backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.2.0"
description = "FastAPI Backend for Wallet Analysis"
authors = ["mateusb12 <matbessa12@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.115.4"
uvicorn = "^0.30.3"
pandas = "^2.3.3"
yfinance = "^0.2.40"
supabase = "^2.4.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
email-validator = "^2.1.0"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"